* update tests * rename `publicKeyHash` to `accountHash` * Add method `CasperClient.getBalanceByAccountHash`
